backend/alembic/versions/0e25765a435f_first_migrations.py
"""first migrations

Revision ID: 0e25765a435f
Revises: 
Create Date: 2024-11-26 21:37:23.164620

"""

from typing import Sequence, Union

from alembic import op
import sqlalchemy as sa
import sqlmodel

# revision identifiers, used by Alembic.
revision: str = "0e25765a435f"
down_revision: Union[str, None] = None
branch_labels: Union[str, Sequence[str], None] = None
depends_on: Union[str, Sequence[str], None] = None


def upgrade() -> None:
    pass


def downgrade() -> None:
    pass
