requirements.txt
setuptools==70.3.0
pywebview
fastapi
uvicorn
pyinstaller; sys_platform != 'darwin'
py2app; sys_platform == 'darwin'
yfinance==0.2.35
ta==0.11.0
langchain-ollama
langchain
sqlmodel
langchain-google-genai
langchain-groq
python-dotenv
pydantic-settings
alembic
